default-autowire-candidates="off1"
even if we have ambigues multiple candidate for the autowiring we can
set the above method for specifing the single coorrespondance.

and this default root level setting can be easily over riden by
introducing the local specific setting to each sub beans.

and multile bean candidates can be specify in the
root like this

default-autowire-candidates="off1,off2"